backend/requirements.txt
fastapi[standard]
uvicorn
pandas
scikit-learn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
google-genai
python-telegram-bot
